+++ crates/cxx-qt-lib-headers/src/lib.rs
@@ -18,6 +18,7 @@ pub fn write_headers(directory: impl AsRef<Path>) {
         (include_str!("../include/common.h"), "common.h"),
         (include_str!("../include/convert.h"), "convert.h"),
         (include_str!("../include/cxxqt_thread.h"), "cxxqt_thread.h"),
+        (include_str!("../include/qbytearray.h"), "qbytearray.h"),
         (include_str!("../include/qcolor.h"), "qcolor.h"),
         (include_str!("../include/qdate.h"), "qdate.h"),
         (include_str!("../include/qdatetime.h"), "qdatetime.h"),